refactor user input checks.
- separate input checks from simulator utils
- adapt tests and usages in inference.
- separate input checks utils
- add Independent Joint with tests.
